+++ backend/src/main/java/codezap/template/dto/response/FindAllTemplatesResponse.java
@@ -4,6 +4,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 import codezap.tag.domain.Tag;
+import codezap.tag.dto.response.FindTagResponse;
 import codezap.template.domain.SourceCode;
 import codezap.template.domain.Template;
 import io.swagger.v3.oas.annotations.media.Schema;
